C programming

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char d;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(d));
-	printf("Size of an int: %lu byte(s)\n", sizeof(a));
-	printf("Size of a long int: %lu bytes(s)\n", sizeof(b));
-	printf("Size of a long long int: %lu bytes(s)\n", sizeof(c));
-	printf("Size of a foat: %lu bytes(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (long unsigned)sizeof(d));
+	printf("Size of an int: %lu byte(s)\n", (long unsigned)sizeof(a));
+	printf("Size of a long int: %lu bytes(s)\n", (long unsigned)sizeof(b));
+	printf("Size of a long long int: %lu bytes(s)\n", (long unsigned)sizeof(c));
+	printf("Size of a foat: %lu bytes(s)\n", (long unsigned)sizeof(f));
 	return (0);
 }
